+++ lib/caching_proxy/cache_entry.rb
@@ -1,4 +1,0 @@
-module CachingProxy
-  class CacheEntry
-  end
-end