+++ lib/caching_proxy/simple_lru_cache.rb
@@ -1,4 +1,0 @@
-module CachingProxy
-  class SimpleLRUCache
-  end
-end