+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require 'rspec'
 require 'caching_proxy'
+require 'timecop'
 
 include CachingProxy