+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ### 0.0.1
 - Initial Release
+- Supports modes for proxying both single and multiple destinations
+- Supports a configurable in-memory cache